Arreglo distribucion de archivos
