+++ COPYPROv2/App.config
@@ -1,0 +1,30 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+    <configSections>
+        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+            <section name="COPYPROv2.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+        </sectionGroup>
+    </configSections>
+    <startup> 
+        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+    </startup>
+    <userSettings>
+        <COPYPROv2.Properties.Settings>
+            <setting name="ColorFondo" serializeAs="String">
+                <value>SteelBlue</value>
+            </setting>
+            <setting name="ColorTexto" serializeAs="String">
+                <value>Snow</value>
+            </setting>
+            <setting name="Fuente" serializeAs="String">
+                <value>Segoe UI Semibold, 9.75pt, style=Bold</value>
+            </setting>
+            <setting name="AnchoForm" serializeAs="String">
+                <value>400</value>
+            </setting>
+            <setting name="AltoForm" serializeAs="String">
+                <value>300</value>
+            </setting>
+        </COPYPROv2.Properties.Settings>
+    </userSettings>
+</configuration>
