+++ COPYPROv2/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="iTextSharp" version="5.5.7" targetFramework="net45" />
+</packages>
